fixed problems with lcm and gcd, added average

+++ src/test/getsimeq.java
@@ -1,8 +1,8 @@
 package test;
 
 public class getsimeq {
-    public static double[] calsimeq(int x1, int y1, int x2, int y2, int res1, int res2) {
-        int det = (x1*y2)-(y1*x2);
+    public static double[] calsimeq(double x1, double y1, double x2, double y2, double res1, double res2) {
+        double det = (x1*y2)-(y1*x2);
         double x = ((y2*res1)+(-y1*res2))/det;
         double y = ((-x2*res1)+(x1*res2))/det;
         double[] result = new double[2];
